Make type conversion checked or assume they are safe
Fix gosec linting on type conversion and overflow. Don’t blindly check
for the overflow, but adapt the code (by sending back the file object
instead of the fd) when needed, or check the underlying type in the API
to ensure we have rooms before overflow.

+++ internal/users/manager.go
@@ -420,8 +420,8 @@ func GenerateID(str string) int {
 	lowerCased := strings.ToLower(str)
 	var sum int
 	for i, c := range lowerCased {
-		// Multiplies the uint value of the rune by its index+1. Subtracts the index to add another layer of conflict prevention.
-		sum += int(uint(c)*uint(i+1)) - i
+		// Multiplies the value of the rune (which is >0) by its index+1. Subtracts the index to add another layer of conflict prevention.
+		sum += int(c)*(i+1) - i
 	}
 	return (sum % (100000 - 65537)) + 65536 // Ensures that ID is between 65536 and 100000
 }
